LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/back-forward-cache/eligibility/inflight-fetch-redirects-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Invalid URL

Harness Error (TIMEOUT), message = null

TIMEOUT Eligibility (in-flight fetch): Redirect header received when in BFCache Test timed out
NOTRUN Eligibility (in-flight fetch): Redirect header received when in BFCache w/ CSP passing
NOTRUN Eligibility (in-flight fetch): Cross-origin redirect header received when in BFCache
NOTRUN Eligibility (in-flight fetch): Cross-origin redirect header received when in BFCache w/ CSP failing
